+++ frontend/src/components/common/Tabs/TabPanel.tsx
@@ -1,0 +1,7 @@
+import { PropsWithChildren } from 'react';
+
+const TabPanel = ({ children }: PropsWithChildren) => {
+  return children;
+};
+
+export default TabPanel;
